extends flock timeout and deals with empty metadata
On the systems I had to test on, ranging from midsized servers to
amazingly underpowered laptops, when using the manta-sync tool,
minimanta would timeout and crash during aquiring file locks.  The
effective default maximum wait was 640ms, due to the check for getting
too long being after the doubling of the previous timeout.  I've
extended it to 10 seconds, and while obviously a huge number, it has
made the operation of minimanta entirely functional for my use cases.

Additionally, the metadata reading occassionally got a 0 length string
and that's not a valid JSON object.  I've added blocks to treat that as
no metadata and act as if it needs default metadata added.